refactor: consolidate dd-register, dd-web, dd-client into unified dd binary
Single `dd` binary with DD_MODE=management (control plane) and
DD_MODE=agent (worker) replaces three separate binaries and container
images. Shared deps mean the unified binary is 9.1MB vs 20.5MB total.

- Convert dd-register, dd-web, dd-client to library crates (pub async fn run())
- New crates/dd dispatches based on DD_MODE env var or CLI arg
- One Dockerfile at repo root replaces three per-crate Dockerfiles
- push-management-images.yml builds one ghcr.io/devopsdefender/dd:latest
- gcp-deploy.sh uses single workload entry with DD_MODE=management
- dd-register defaults to port 8081 via DD_REGISTER_PORT
